structure change

+++ run.ps1
@@ -3,15 +3,18 @@ $resourceGroupName='test'
 
 if (az group exists --resource-group $resourceGroupName){
 	echo 'Resource group already exists.'
-	try{
-		while ($true){
-			az resource delete --name (az resource list --query '[0].{name:name}' --resource-group test --output table)[2] --resource-type (az resource list --query [0]'.{name:type}' --resource-group test --output table)[2] --resource-group test
-			Start-Sleep -Seconds 1.5
-		}	
-	}
-	catch{
-		echo 'Resources removed.'
-		az ml workspace create --resource-group test --name testazml --location eastus
+	if (False){
+		try{
+			while ($true){
+				az resource delete --name (az resource list --query '[0].{name:name}' --resource-group test --output table)[2] --resource-type (az resource list --query [0]'.{name:type}' --resource-group test --output table)[2] --resource-group test
+				Start-Sleep -Seconds 1.5
+			}		
+		}
+		catch{
+			echo 'Resources removed.'
+			echo 'Creating a new workspace...'
+			az ml workspace create --resource-group test --name testazml --location eastus
+		}
 	}
 }
 else{
@@ -19,7 +22,10 @@ else{
 	az group create --name $resourceGroupName --location eastus
 	az ml workspace create --resource-group test --name testazml --location eastus
 }
+echo 'Extraction and transformation. Running...'
+$storageBlob_conn = (az storage account show-connection-string --name testsa --resource-group test --query 'connectionString' --output tsv)
+python ./run_workflow.py testsa, $storageBlob_conn, test
 $storageName = (az ml datastore list --query '[0].{storageName:account_name}' --output table --workspace-name testazml --resource-group test)[2]
 $storageKey = (az storage account keys list --resource-group test --account-name $storageName --query '[0].value' --output tsv)
-az storage blob upload --account-name $storageName --account-key $storageKey --container-name azureml --file ./run_workflow.py --name run_workflow.py --overwrite
+az storage blob upload --account-name $storageName --account-key $storageKey --container-name azureml --file ./jobs/job_workflow.py --name job_workflow.py --overwrite
 az ml job create --name test_pipeline --file ./jobs/experiment_job.yml --workspace-name testazml --resource-group test
